feat: support streaming responses in the component handler
