+++ runtime_test.go
@@ -377,102 +377,6 @@ func TestClassesFunction(t *testing.T) {
 	}
 }
 
-func TestHandler(t *testing.T) {
-	hello := templ.ComponentFunc(func(ctx context.Context, w io.Writer) error {
-		if _, err := io.WriteString(w, "Hello"); err != nil {
-			t.Fatalf("failed to write string: %v", err)
-		}
-		return nil
-	})
-	errorComponent := templ.ComponentFunc(func(ctx context.Context, w io.Writer) error {
-		if _, err := io.WriteString(w, "Hello"); err != nil {
-			t.Fatalf("failed to write string: %v", err)
-		}
-		return errors.New("handler error")
-	})
-
-	tests := []struct {
-		name             string
-		input            *templ.ComponentHandler
-		expectedStatus   int
-		expectedMIMEType string
-		expectedBody     string
-	}{
-		{
-			name:             "handlers return OK by default",
-			input:            templ.Handler(hello),
-			expectedStatus:   http.StatusOK,
-			expectedMIMEType: "text/html; charset=utf-8",
-			expectedBody:     "Hello",
-		},
-		{
-			name:             "handlers return OK by default",
-			input:            templ.Handler(templ.Raw(`♠ ‘ &spades; &#8216;`)),
-			expectedStatus:   http.StatusOK,
-			expectedMIMEType: "text/html; charset=utf-8",
-			expectedBody:     "♠ ‘ &spades; &#8216;",
-		},
-		{
-			name:             "handlers can be configured to return an alternative status code",
-			input:            templ.Handler(hello, templ.WithStatus(http.StatusNotFound)),
-			expectedStatus:   http.StatusNotFound,
-			expectedMIMEType: "text/html; charset=utf-8",
-			expectedBody:     "Hello",
-		},
-		{
-			name:             "handlers can be configured to return an alternative status code and content type",
-			input:            templ.Handler(hello, templ.WithStatus(http.StatusOK), templ.WithContentType("text/csv")),
-			expectedStatus:   http.StatusOK,
-			expectedMIMEType: "text/csv",
-			expectedBody:     "Hello",
-		},
-		{
-			name:             "handlers that fail return a 500 error",
-			input:            templ.Handler(errorComponent),
-			expectedStatus:   http.StatusInternalServerError,
-			expectedMIMEType: "text/plain; charset=utf-8",
-			expectedBody:     "templ: failed to render template\n",
-		},
-		{
-			name: "error handling can be customised",
-			input: templ.Handler(errorComponent, templ.WithErrorHandler(func(r *http.Request, err error) http.Handler {
-				// Because the error is received, it's possible to log the detail of the request.
-				// log.Printf("template render error for %v %v: %v", r.Method, r.URL.String(), err)
-				return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-					w.WriteHeader(http.StatusBadRequest)
-					if _, err := io.WriteString(w, "custom body"); err != nil {
-						t.Fatalf("failed to write string: %v", err)
-					}
-				})
-			})),
-			expectedStatus:   http.StatusBadRequest,
-			expectedMIMEType: "text/html; charset=utf-8",
-			expectedBody:     "custom body",
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			w := httptest.NewRecorder()
-			r := httptest.NewRequest("GET", "/test", nil)
-			tt.input.ServeHTTP(w, r)
-			if got := w.Result().StatusCode; tt.expectedStatus != got {
-				t.Errorf("expected status %d, got %d", tt.expectedStatus, got)
-			}
-			if mimeType := w.Result().Header.Get("Content-Type"); tt.expectedMIMEType != mimeType {
-				t.Errorf("expected content-type %s, got %s", tt.expectedMIMEType, mimeType)
-			}
-			body, err := io.ReadAll(w.Result().Body)
-			if err != nil {
-				t.Errorf("failed to read body: %v", err)
-			}
-			if diff := cmp.Diff(tt.expectedBody, string(body)); diff != "" {
-				t.Error(diff)
-			}
-		})
-	}
-}
-
 func TestRenderScriptItems(t *testing.T) {
 	s1 := templ.ComponentScript{
 		Name:     "s1",
